Deepen query and runtime modules
