+++ packages/column-live-view-engine/src/live-subscription.ts
@@ -1,11 +1,7 @@
-import type { DeltaEvent, DeltaOperation, SnapshotEvent, StatusEvent } from "@view-server/config";
+import type { DeltaEvent, SnapshotEvent, StatusEvent } from "@view-server/config";
 import { Cause, Effect, Queue, Stream } from "effect";
-import {
-  evaluateCompiledRawQuery,
-  type CompiledRawQuery,
-  type QueryEvaluation,
-} from "./raw-query-compiler";
-import { cloneRow, rowsEqual } from "./row-values";
+import { evaluateCompiledRawQuery, type CompiledRawQuery } from "./raw-query-compiler";
+import { deltaEvent, deltaOperations, snapshotEvent } from "./query-result";
 
 type RowObject = object;
 
@@ -45,113 +41,6 @@ export type LiveSubscription<ResultRow extends RowObject> = {
   readonly events: Stream.Stream<LiveSubscriptionEvent<ResultRow>>;
   readonly close: () => Effect.Effect<void, never>;
 };
-
-const snapshotEvent = <Row extends RowObject>(
-  store: { readonly topic: string },
-  queryId: string,
-  evaluation: QueryEvaluation<Row>,
-): SnapshotEvent<Row> => ({
-  type: "snapshot",
-  topic: store.topic,
-  queryId,
-  version: evaluation.version,
-  keys: [...evaluation.keys],
-  rows: evaluation.rows.map(cloneRow),
-  totalRows: evaluation.totalRows,
-});
-
-const deltaOperations = <Row extends RowObject>(
-  previous: QueryEvaluation<Row>,
-  next: QueryEvaluation<Row>,
-): ReadonlyArray<DeltaOperation<Row>> => {
-  const operations: Array<DeltaOperation<Row>> = [];
-  const nextKeys = new Set(next.keys);
-  const currentKeys = [...previous.keys];
-  const currentRows = [...previous.rows];
-
-  for (const key of previous.keys) {
-    if (!nextKeys.has(key)) {
-      const index = currentKeys.indexOf(key);
-      currentKeys.splice(index, 1);
-      currentRows.splice(index, 1);
-      operations.push({
-        type: "remove",
-        key,
-      });
-    }
-  }
-
-  for (const [index, { key, row }] of next.window.entries()) {
-    const currentIndex = currentKeys.indexOf(key);
-    if (currentIndex < 0) {
-      currentKeys.splice(index, 0, key);
-      currentRows.splice(index, 0, row);
-      operations.push({
-        type: "insert",
-        key,
-        row,
-        index,
-      });
-      continue;
-    }
-
-    if (currentIndex !== index) {
-      const currentRow = currentRows[currentIndex]!;
-      currentKeys.splice(currentIndex, 1);
-      currentRows.splice(currentIndex, 1);
-      currentKeys.splice(index, 0, key);
-      currentRows.splice(index, 0, currentRow);
-      operations.push({
-        type: "move",
-        key,
-        fromIndex: currentIndex,
-        toIndex: index,
-      });
-    }
-
-    const currentRow = currentRows[index];
-    if (currentRow === undefined || !rowsEqual(currentRow, row)) {
-      currentRows[index] = row;
-      operations.push({
-        type: "update",
-        key,
-        row,
-        index,
-      });
-    }
-  }
-
-  return operations;
-};
-
-const cloneDeltaOperations = <Row extends RowObject>(
-  operations: ReadonlyArray<DeltaOperation<Row>>,
-): ReadonlyArray<DeltaOperation<Row>> =>
-  operations.map((operation) => {
-    if (operation.type === "insert" || operation.type === "update") {
-      return {
-        ...operation,
-        row: cloneRow(operation.row),
-      };
-    }
-    return operation;
-  });
-
-const deltaEvent = <Row extends RowObject>(
-  store: { readonly topic: string },
-  queryId: string,
-  fromVersion: number,
-  next: QueryEvaluation<Row>,
-  operations: ReadonlyArray<DeltaOperation<Row>>,
-): DeltaEvent<Row> => ({
-  type: "delta",
-  topic: store.topic,
-  queryId,
-  fromVersion,
-  toVersion: next.version,
-  operations: cloneDeltaOperations(operations),
-  totalRows: next.totalRows,
-});
 
 const backpressureStatusEvent = (
   store: { readonly topic: string },
