+++ packages/config/src/query-exact.ts
@@ -1,0 +1,3 @@
+export type RejectExtraKeys<Candidate, Shape> = {
+  readonly [Key in Exclude<keyof Candidate, keyof Shape>]: never;
+};
